Adding "is empty" and "is not empty" comparisons
This adds "is empty" and "is not empty" unary comparisons to hyperscript.  It looks like these are the only unary expressions in the grammar, so I added a flag into this function (true by default) so that operators can choose to require a right hand side value or not.

This commit includes tests (all passing) and additions to the documentation.